refractoring

+++ CRM4telecom/CRM4telecom-ejb/src/java/com/crm4telecom/jpa/MarketProducts.java
@@ -1,36 +1,23 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 package com.crm4telecom.jpa;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.util.Date;
 import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
-import javax.xml.bind.annotation.XmlRootElement;
 
-/**
- *
- * @author Alex
- */
 @Entity
 @Table(catalog = "", schema = "CRM4TELECOM")
 public class MarketProducts implements Serializable {
+
     private static final long serialVersionUID = 1L;
     @EmbeddedId
     protected MarketProductsPK marketProductsPK;
     @Column(name = "PRICE")
-    private BigInteger price;
+    private Long price;
     @Column(name = "START_DATE")
     @Temporal(TemporalType.TIMESTAMP)
     private Date startDate;
@@ -45,7 +32,7 @@ public class MarketProducts implements Serializable {
         this.marketProductsPK = marketProductsPK;
     }
 
-    public MarketProducts(BigInteger marketId, BigInteger productId) {
+    public MarketProducts(Long marketId, Long productId) {
         this.marketProductsPK = new MarketProductsPK(marketId, productId);
     }
 
@@ -57,11 +44,11 @@ public class MarketProducts implements Serializable {
         this.marketProductsPK = marketProductsPK;
     }
 
-    public BigInteger getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(BigInteger price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
 
@@ -105,5 +92,5 @@ public class MarketProducts implements Serializable {
     public String toString() {
         return "com.crm4telecom.jpa.MarketProducts[ marketProductsPK=" + marketProductsPK + " ]";
     }
-    
+
 }
